pyproject.toml
[project]
name = "outomem"
version = "0.1.0"
description = "The ultimate memory system library for AI agents."
requires-python = ">=3.10"
dependencies = [
    "lancedb>=0.5.0",
    "fastembed>=0.3.0",
    "anthropic>=0.20.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.5.0",
    "python-dateutil>=2.8.0",
    "neo4j>=5.0.0",
    "pyarrow>=14.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
]
